Create ApiException.php
ApiException class allows us to create a custom error message. The body of the message can be a string or object. Also the class allows us to set the Http code.

This will save a lot of code when someone writes a new controller and methods.

Please take into account that this change is connected to a change proposed for APIJet.php method run()!